+++ src/plugin.ts
@@ -14,6 +14,7 @@ import './input-field-dialog'
 import './object-slider-dialog'
 import './memory-leak-fix'
 import './pvp-damage-factor-override'
+import './mute-map-sounds'
 
 export default class KrypekLib implements PluginClass {
     static dir: string